requirements.txt
markdown==3.5.2
pymdown-extensions==10.8
PyYAML==6.0.1 